PI-16960 Moved to service folder
