+++ src/main/java/com/appdirect/sdk/security/encryption/service/EncryptionService.java
@@ -1,4 +1,6 @@
-package com.appdirect.sdk.security.encryption;
+package com.appdirect.sdk.security.encryption.service;
+
+import com.appdirect.sdk.security.encryption.EncryptionException;
 
 public interface EncryptionService {
 	String encrypt(String value) throws EncryptionException;
